Switch float and double 🤦‍♂️Remove exp in decimals

+++ ellie_engine/bytecode/src/addressing_modes.rs
@@ -130,10 +130,10 @@ impl core::fmt::Display for AddressingModes {
                         }
                         .to_string(),
                         Types::Float =>
-                            f64::from_le_bytes(value[0..mem::size_of::<f64>()].try_into().unwrap())
+                            f32::from_le_bytes(value[0..mem::size_of::<f32>()].try_into().unwrap())
                                 .to_string(),
                         Types::Double =>
-                            f32::from_le_bytes(value[0..mem::size_of::<f32>()].try_into().unwrap())
+                            f64::from_le_bytes(value[0..mem::size_of::<f64>()].try_into().unwrap())
                                 .to_string(),
                         Types::Byte => format!("0x{}", value[0]),
                         Types::Bool =>
